Extract class to fixtures.
