+++ tests/Fixtures/Bar.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tests\Fixtures;
+
+use Illuminate\Database\Eloquent\Model;
+use Sushi\Sushi;
+
+class Bar extends Model
+{
+    use Sushi;
+}
